+++ packages/marvin-frontend-new/src/app/components/GraphStyles.scss
@@ -1,6 +1,51 @@
 svg {
     circle.hover {
         filter: drop-shadow(0 0 4px rgb(0 0 0 / 0.6));
-        fill: rgb(255,255,255);
+        fill: rgb(255, 255, 255);
     }
+
+    .node.disabled {
+        opacity: 0.1;
+    }
+
+    .node.loading {
+        animation: fadeInOut 3s ease-in-out 0s infinite normal forwards;
+    }
+
+   
 }
+
+.backdrop {
+    position: absolute;
+    background: rgba(0, 0, 0, 0.8);
+    z-index: 9999;
+    left: 0;
+    right: 0;
+    top: 0;
+    bottom: 0;
+}
+
+.coverimage {
+    position: absolute;
+    background: #000;
+    z-index: 1;
+    max-height: 90%;
+    left: 50%;
+    margin-left: -25%;
+    top: 70px;
+    border: 2px solid #000;
+}
+
+
+@keyframes fadeInOut {
+    0%,
+	50%,
+	100% {
+		opacity: 0.5;
+	}
+
+	25%,
+	75% {
+		opacity: 0.2;
+	}
+  }
