+++ packages/marvin-frontend-new/src/styles.scss
@@ -11,6 +11,7 @@ body {
   padding: 0;
   padding-top: $pt-navbar-height;
   margin: 0;
+  background-color: #252a31;
 }
 
 html, body {
